pyproject.toml
[project]
name = "stripe-des"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "stripe>=14.0.1",
    "python-dotenv>=1.0.0",
]
